requirements.txt
gspread
pygsheets
numpy
openpyxl
pandas
polars
plotly_express
tables